# swagger.yml
openapi: 3.0.0
info:
  description: devops hometask
  version: 1.0.2
  title: Task Swagger
  termsOfService: http://swagger.io/terms/
  contact:
    email: avborovets@gmail.com
tags:
  - name: Users
    description: Everything about Users in db
    externalDocs:
      description: Find out more
      url: http://swagger.io

paths:
  /users:
    get:
      tags:
        - Users
      summary: Get users from db
      operationId: getUsers
      responses:
        '200':
          description: Users successfully are got
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
        'default':
          description: Error!!!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  
    post:
      tags:
        - Users
      summary: Add a new user to the db
      description: ""
      operationId: addUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '200':
          description: User is successfully created
        'default':
          description: default message

  /users/{id}:
    put:
      tags:
        - Users
      summary: Update an existing user
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      parameters:
        - name: id
          in: path
          required: true
          description: user id
          schema:
            type: string
      responses:
        '200':
          description:  User is updated successfully
        'default':
          description: Error!!!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Users
      summary: Delete an existing user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: user id
          schema:
            type: string
      responses:
        '200':
          description:  User is deleted successfully
        'default':
          description: Error!!!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          

externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io

servers:
  - url: http://127.0.0.1:8099


components:
  schemas:
    User:
      type: object
      required:
        - nickname
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        email:
          type: string
        rating:
          type: integer
          format: int64
    Users:
      type: array
      items:
        $ref: "#/components/schemas/User"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
 